refactored eval run selection logic out to components/
