+++ ui/app/components/evaluations/ColorAssigner.tsx
